AdminSubCategory.java

+++ src/main/java/com/chiasetailieu/dao/ISubcategoryDAO.java
@@ -10,7 +10,7 @@ public interface ISubcategoryDAO extends IGenericDAO<SubCategory>{
 	List<SubCategory> findAll();
 	List<SubCategory> findByCategories(Category category);
 	List<SubCategory> findOneByDocument(Document document);
-	SubCategory findOneById(int id);
+	SubCategory findOneById(long id);
 	Long save(SubCategory subcate);
 	void update(SubCategory subcate);
 	void delete(SubCategory subcate);
